cmd/snap-preseed: move the sources around
Move the sources around in preparation for merging of snap-preseed into
snapd.

Related: SNAPDENG-37053
