+++ cmd/snapd/tool/snap-preseed/preseed_uc20_test.go
@@ -17,7 +17,7 @@
  *
  */
 
-package main_test
+package snap_preseed_test
 
 import (
 	"os"
@@ -25,7 +25,7 @@ import (
 
 	. "gopkg.in/check.v1"
 
-	"github.com/snapcore/snapd/cmd/snap-preseed"
+	"github.com/snapcore/snapd/cmd/snapd/tool/snap-preseed"
 	"github.com/snapcore/snapd/dirs"
 	"github.com/snapcore/snapd/image/preseed"
 )
@@ -34,7 +34,7 @@ func (s *startPreseedSuite) TestRunPreseedUC20Happy(c *C) {
 	tmpDir := c.MkDir()
 	dirs.SetRootDir(tmpDir)
 
-	restore := main.MockOsGetuid(func() int {
+	restore := snap_preseed.MockOsGetuid(func() int {
 		return 0
 	})
 	defer restore()
@@ -45,7 +45,7 @@ func (s *startPreseedSuite) TestRunPreseedUC20Happy(c *C) {
 	c.Assert(os.WriteFile(filepath.Join(tmpDir, "system-seed/systems/20220203/preseed.tgz"), nil, 0644), IsNil)
 
 	var called bool
-	restorePreseed := main.MockPreseedCore20(func(opts *preseed.CoreOptions) error {
+	restorePreseed := snap_preseed.MockPreseedCore20(func(opts *preseed.CoreOptions) error {
 		c.Check(opts.PrepareImageDir, Equals, tmpDir)
 		c.Check(opts.PreseedSignKey, Equals, "key")
 		c.Check(opts.AppArmorKernelFeaturesDir, Equals, "/custom/aa/features")
@@ -56,7 +56,7 @@ func (s *startPreseedSuite) TestRunPreseedUC20Happy(c *C) {
 	defer restorePreseed()
 
 	parser := testParser(c)
-	c.Assert(main.Run(parser, []string{"--preseed-sign-key", "key", "--apparmor-features-dir", "/custom/aa/features", "--sysfs-overlay", "/sysfs-overlay", tmpDir}), IsNil)
+	c.Assert(snap_preseed.Run(parser, []string{"--preseed-sign-key", "key", "--apparmor-features-dir", "/custom/aa/features", "--sysfs-overlay", "/sysfs-overlay", tmpDir}), IsNil)
 	c.Check(called, Equals, true)
 }
 
@@ -64,7 +64,7 @@ func (s *startPreseedSuite) TestRunPreseedUC20HappyNoArgs(c *C) {
 	tmpDir := c.MkDir()
 	dirs.SetRootDir(tmpDir)
 
-	restore := main.MockOsGetuid(func() int {
+	restore := snap_preseed.MockOsGetuid(func() int {
 		return 0
 	})
 	defer restore()
@@ -74,7 +74,7 @@ func (s *startPreseedSuite) TestRunPreseedUC20HappyNoArgs(c *C) {
 	c.Assert(os.WriteFile(filepath.Join(tmpDir, "system-seed/systems/20220203/preseed.tgz"), nil, 0644), IsNil)
 
 	var called bool
-	restorePreseed := main.MockPreseedCore20(func(opts *preseed.CoreOptions) error {
+	restorePreseed := snap_preseed.MockPreseedCore20(func(opts *preseed.CoreOptions) error {
 		c.Check(opts.PrepareImageDir, Equals, tmpDir)
 		c.Check(opts.PreseedSignKey, Equals, "")
 		c.Check(opts.AppArmorKernelFeaturesDir, Equals, "")
@@ -85,7 +85,7 @@ func (s *startPreseedSuite) TestRunPreseedUC20HappyNoArgs(c *C) {
 	defer restorePreseed()
 
 	parser := testParser(c)
-	c.Assert(main.Run(parser, []string{tmpDir}), IsNil)
+	c.Assert(snap_preseed.Run(parser, []string{tmpDir}), IsNil)
 	c.Check(called, Equals, true)
 }
 
@@ -93,7 +93,7 @@ func (s *startPreseedSuite) TestResetUC20(c *C) {
 	tmpDir := c.MkDir()
 	dirs.SetRootDir(tmpDir)
 
-	restore := main.MockOsGetuid(func() int {
+	restore := snap_preseed.MockOsGetuid(func() int {
 		return 0
 	})
 	defer restore()
@@ -104,14 +104,14 @@ func (s *startPreseedSuite) TestResetUC20(c *C) {
 	c.Assert(os.WriteFile(filepath.Join(tmpDir, "system-seed/systems/20220203/preseed.tgz"), nil, 0644), IsNil)
 
 	var called bool
-	restorePreseed := main.MockPreseedCore20(func(opts *preseed.CoreOptions) error {
+	restorePreseed := snap_preseed.MockPreseedCore20(func(opts *preseed.CoreOptions) error {
 		called = true
 		return nil
 	})
 	defer restorePreseed()
 
 	parser := testParser(c)
-	res := main.Run(parser, []string{"--reset", tmpDir})
+	res := snap_preseed.Run(parser, []string{"--reset", tmpDir})
 	c.Assert(res, Not(IsNil))
 	c.Check(res, ErrorMatches, "cannot snap-preseed --reset for Ubuntu Core")
 	c.Check(called, Equals, false)
